aply times to laboratory orders (time of take, delivery and accept)

+++ database/migrations/2022_12_22_212601_create_laboratory_orders_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
           $table->string('lo_name');
           $table->string('lo_name_en');
           $table->boolean('lo_break')->default(0);
-          $table->tinyInteger('lo_sort');
+          $table->Integer('lo_sort');
           $table->timestamps();
       });
       Schema::table('laboratory_orders', function (Blueprint $table) {
